// src/alice/types/yandex.ts
export enum YDevice {
	camera = "camera",
	cooking = "cooking",
	cooking_coffee_maker = "cooking.coffee_maker",
	cooking_kettle = "cooking.kettle",
	cooking_multicooker = "cooking.multicooker",
	dishwasher = "dishwasher",
	humidifier = "humidifier",
	iron = "iron",
	light = "light",
	light_ceiling = "light.ceiling",
	light_strip = "light.strip",
	media_device = "media_device",
	media_device_receiver = "media_device.receiver",
	media_device_tv = "media_device.tv",
	media_device_tv_box = "media_device.tv_box",
	openable = "openable",
	openable_curtain = "openable.curtain",
	openable_valve = "openable.valve",
	other = "other",
	pet_drinking_fountain = "pet_drinking_fountain",
	pet_feeder = "pet_feeder",
	purifier = "purifier",
	sensor = "sensor",
	sensor_button = "sensor.button",
	sensor_climate = "sensor.climate",
	sensor_gas = "sensor.gas",
	sensor_illumination = "sensor.illumination",
	sensor_motion = "sensor.motion",
	sensor_open = "sensor.open",
	sensor_smoke = "sensor.smoke",
	sensor_vibration = "sensor.vibration",
	sensor_water_leak = "sensor.water_leak",
	smart_meter = "smart_meter",
	smart_meter_cold_water = "smart_meter.cold_water",
	smart_meter_electricity = "smart_meter.electricity",
	smart_meter_gas = "smart_meter.gas",
	smart_meter_heat = "smart_meter.heat",
	smart_meter_hot_water = "smart_meter.hot_water",
	socket = "socket",
	switch = "switch",
	thermostat = "thermostat",
	thermostat_ac = "thermostat.ac",
	vacuum_cleaner = "vacuum_cleaner",
	ventilation = "ventilation",
	ventilation_fan = "ventilation.fan",
	washing_machine = "washing_machine",
}